requirements.txt
openpyxl=3.1.5
SQLAlchemy=2.0.36
pandas=2.2.3
dash=2.18.2
python-dotenv=1.0.1